Add files via upload
Here when you will run the code a pop up will be seen where you can register or login if you already have an account. Another window will be poped  and you will be asked either to create note, view note, or delete note and you can carry out that operations as same. This project was just intantiated for learning purpose of tkinter library of python